// src/app/core/models/report.interface.ts
export interface Report {
    representanteComercial: string;
    nombre: string;
    codpanel: string;
    idComunicador: string;
    imei: string;
    contrato: string;
    numSerie: string;
    estado: string;
    habilitada: string;
    fechaHabilitacion: string;
    fechaDeshabilitacion: string;
    modelo: string;
    numTelefonoSim: string;
    numImsi: string;
    exton: string;
    canalSms: string;
    canalGprs: string;
    smson: string;
    be: string;
    apn: string;
}